+++ src/pdm_conda/conda.py
@@ -85,18 +85,25 @@ def run_conda(
         logger.debug(f"cmd: {' '.join(cmd)}")
         if environment:
             logger.debug(f"env: {environment}")
-        process = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding="utf-8")
+        process = subprocess.run(cmd, capture_output=True, encoding="utf-8")
     if "--json" in cmd:
         try:
             response = json.loads(process.stdout)
         except:
             response = {}
     else:
-        response = {"message": f"{process.stdout}\n{process.stderr}"}
+        response = {}
+        msg = f"{process.stdout}\n" if process.stdout else ""
+        if process.stderr:
+            msg += process.stderr
+        if msg:
+            response["message"] = msg
     try:
         process.check_returncode()
     except subprocess.CalledProcessError as e:
-        msg = f"{exception_msg}\n"
+        msg = ""
+        if exception_msg:
+            msg = f"{exception_msg}\n"
         if isinstance(response, dict) and not response.get("success", False):
             if err := response.get("solver_problems", response.get("error", response.get("message", process.stderr))):
                 if isinstance(err, str):
@@ -245,7 +252,7 @@ def conda_search(
     if "::" in _requirement:
         channel, _requirement = _requirement.split("::", maxsplit=1)
     if isinstance(requirement, str):
-        requirement = parse_requirement(f"conda::{requirement}")
+        requirement = parse_requirement(f"conda:{requirement}")
     channels = _ensure_channels(
         project,
         [channel] if channel else [],
@@ -360,7 +367,7 @@ def _conda_install(
     kwargs: dict = dict()
     if explicit:
         kwargs["lockfile"] = packages
-    run_conda(command + ["--json"], **kwargs)
+    run_conda(command + ["--json"], exception_cls=exception_cls, exception_msg="", **kwargs)
 
 
 def conda_install(
@@ -407,11 +414,7 @@ def conda_uninstall(
     """
     config = project.conda_config
 
-    with config.with_config(
-        runner=CondaRunner.CONDA if no_deps and config.runner == CondaRunner.MAMBA else config.runner,
-    ):
-        command = config.command("remove")
-
+    command = config.command("remove")
     if no_deps:
         if config.runner == CondaRunner.MICROMAMBA:
             command.append("--no-prune")
@@ -460,6 +463,8 @@ def conda_list(project: CondaProject) -> dict[str, CondaSetupDistribution]:
     if config.is_initialized:
         packages = run_conda(config.command("list") + ["--json"])
         for package in packages:
+            if config.runner != CondaRunner.MICROMAMBA and package.get("platform", "") == "pypi":
+                continue
             name, version = package["name"], package["version"]
             distributions[normalize_name(name)] = CondaSetupDistribution(
                 Setup(