+++ src/pdm_conda/models/environment.py
@@ -1,6 +1,7 @@
 import functools
 import os
 import sysconfig
+import uuid
 from copy import copy
 from pathlib import Path
 from typing import cast
@@ -12,8 +13,9 @@ from pdm.models.specifiers import PySpecSet
 from pdm.models.working_set import WorkingSet
 from pdm.project import Project
 
-from pdm_conda.conda import conda_list, conda_search
+from pdm_conda.conda import conda_create, conda_list, conda_search
 from pdm_conda.mapping import pypi_to_conda
+from pdm_conda.models.config import CondaRunner, CondaSolver
 from pdm_conda.project import CondaProject
 from pdm_conda.utils import normalize_name
 
@@ -30,7 +32,7 @@ class CondaEnvironment(Environment):
     def __init__(self, project: Project) -> None:
         super().__init__(project)
         self.project = cast(CondaProject, project)
-        self._python_dependencies: dict[str, Requirement] | None = None
+        self._env_dependencies: dict[str, Requirement] | None = None
         self.python_requires &= PySpecSet(f"=={self.interpreter.version}")
 
     @property
@@ -54,22 +56,37 @@ class CondaEnvironment(Environment):
         return working_set
 
     @property
-    def python_dependencies(self) -> dict[str, Requirement]:
-        if self._python_dependencies is None:
-            self._python_dependencies = dict()
+    def env_dependencies(self) -> dict[str, Requirement]:
+        if self._env_dependencies is None:
+            self._env_dependencies = dict()
 
             def load_dependencies(name: str, packages: dict, dependencies: dict):
                 if name not in packages and name not in dependencies:
                     return
-                package = packages[name].as_line().replace(" ", "=")
-                candidate = conda_search(self.project, package)[0]
+                candidate = conda_search(self.project, packages[name].req)[0]
                 dependencies[name] = candidate.req
                 for d in candidate.dependencies:
                     load_dependencies(d.name, packages, dependencies)
 
-            load_dependencies("python", conda_list(self.project), self._python_dependencies)
+            working_set = conda_list(self.project)
+            dependencies = ["python"]
+            if (runner := self.project.conda_config.runner) in working_set:
+                dependencies.append(runner)
+            if (
+                runner in (CondaRunner.MAMBA, CondaRunner.MICROMAMBA)
+                or self.project.conda_config.solver == CondaSolver.MAMBA
+            ):
+                self._env_dependencies = conda_create(
+                    self.project,
+                    [working_set[d].req for d in dependencies],
+                    prefix=f"/tmp/{uuid.uuid4()}",
+                    dry_run=True,
+                )
+            else:
+                for dep in dependencies:
+                    load_dependencies(dep, working_set, self._env_dependencies)
 
-        return self._python_dependencies
+        return self._env_dependencies
 
 
 def wrap_init(func):