+++ tests/test_mapping.py
@@ -13,15 +13,33 @@ def patch_download_dir(monkeypatch):
         yield d
 
 
+@pytest.fixture
+def patch_conda_mapping_fixes(mocker, conda_mapping_fixes):
+    yield mocker.patch("pdm_conda.mapping.get_mapping_fixes", return_value=conda_mapping_fixes)
+
+
 class TestMapping:
     @pytest.mark.parametrize("conda_mapping", [{"pytest": "pytest-conda", "other": "other-conda"}, dict()])
-    def test_download_mapping(self, patch_download_dir, project, conda_mapping, mocked_responses):
+    @pytest.mark.parametrize(
+        "conda_mapping_fixes",
+        [{"corrected-mapping": "corrected"}, {"pytest": "pytest-corrected"}],
+    )
+    def test_download_mapping(
+        self,
+        patch_download_dir,
+        project,
+        conda_mapping,
+        patch_conda_mapping_fixes,
+        conda_mapping_fixes,
+        mocked_responses,
+    ):
         """
         Test project conda_mapping downloads conda mapping just one and mapping is as expected
         """
-        from pdm_conda.mapping import MAPPINGS_URL, get_pypi_mapping
+        from pdm_conda.mapping import MAPPINGS_URL, get_conda_mapping, get_pypi_mapping
 
         get_pypi_mapping.cache_clear()
+        get_conda_mapping.cache_clear()
 
         assert project.config["conda.pypi-mapping.download-dir"] == patch_download_dir
 
@@ -36,16 +54,37 @@ class TestMapping:
             """
         rsp = mocked_responses.get(MAPPINGS_URL, body=response)
 
+        corrected_mapping = dict(conda_mapping)
+        corrected_mapping.update(conda_mapping_fixes)
         for _ in range(5):
-            assert get_pypi_mapping() == conda_mapping
+            assert get_pypi_mapping() == corrected_mapping
+            assert get_conda_mapping() == {k: v for v, k in corrected_mapping.items()}
         assert rsp.call_count == 1
+        assert patch_conda_mapping_fixes.call_count == 1
         for ext in ["yaml", "json"]:
             assert (Path(patch_download_dir) / f"pypi_mapping.{ext}").exists()
 
     @pytest.mark.parametrize("conda_mapping", [{"pytest": "pytest-conda", "other": "other-conda"}])
     @pytest.mark.parametrize("package", ["pytest", "other", "otherPackage"])
-    def test_pypi_mapping(self, patch_download_dir, project, conda_mapping, package, mocked_responses):
-        self.test_download_mapping(patch_download_dir, project, conda_mapping, mocked_responses)
+    @pytest.mark.parametrize("conda_mapping_fixes", [dict()])
+    def test_pypi_mapping(
+        self,
+        patch_download_dir,
+        project,
+        conda_mapping,
+        patch_conda_mapping_fixes,
+        conda_mapping_fixes,
+        package,
+        mocked_responses,
+    ):
+        self.test_download_mapping(
+            patch_download_dir,
+            project,
+            conda_mapping,
+            patch_conda_mapping_fixes,
+            conda_mapping_fixes,
+            mocked_responses,
+        )
         from pdm_conda.mapping import pypi_to_conda
 
         assert pypi_to_conda(package) == conda_mapping.get(package, package.lower())